feat: remove unused Dropdown components and styles
