--- src/components/LexicalFocused/Editor.css
@@ -1,34 +1,0 @@
-.editor-shell {
-  position: relative;
-  display: block;
-  background: #fff;
-  color: #000;
-  line-height: 1.7;
-  font-weight: 400;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  border: 1px solid #eee;
-  border-radius: 10px;
-  box-shadow:
-    1px 1px 3px #eee,
-    -1px -1px 3px #eee;
-}
-
-.editor-scroller {
-  min-height: 150px;
-  max-width: 100%;
-  border: 0;
-  display: flex;
-  position: relative;
-  outline: 0;
-  z-index: 0;
-  resize: vertical;
-}
-
-.editor {
-  flex: auto;
-  max-width: 100%;
-  position: relative;
-  resize: vertical;
-  z-index: -1;
-}
