--- src/components/LexicalFocused/components/Dropdown/Dropdown.css
@@ -1,135 +1,0 @@
-.dropdown {
-  z-index: 100;
-  display: block;
-  position: fixed;
-  box-shadow:
-    0 12px 28px 0 rgba(0, 0, 0, 0.2),
-    0 2px 4px 0 rgba(0, 0, 0, 0.1),
-    inset 0 0 0 1px rgba(255, 255, 255, 0.5);
-  border-radius: 8px;
-  min-height: 40px;
-  background-color: #fff;
-}
-
-.dropdown .item {
-  margin: 0 8px 0 8px;
-  padding: 8px;
-  color: #050505;
-  cursor: pointer;
-  line-height: 16px;
-  font-size: 15px;
-  display: flex;
-  align-content: center;
-  flex-direction: row;
-  flex-shrink: 0;
-  justify-content: space-between;
-  background-color: #fff;
-  border-radius: 8px;
-  border: 0;
-  max-width: 264px;
-  min-width: 100px;
-}
-
-.dropdown .item.wide {
-  align-items: center;
-  width: 260px;
-}
-
-.dropdown .item.wide .icon-text-container {
-  display: flex;
-
-  .text {
-    min-width: 120px;
-  }
-}
-
-.dropdown .item .shortcut {
-  color: #939393;
-  align-self: flex-end;
-}
-
-.dropdown .item .active {
-  display: flex;
-  width: 20px;
-  height: 20px;
-  background-size: contain;
-}
-
-.dropdown .item:first-child {
-  margin-top: 8px;
-}
-
-.dropdown .item:last-child {
-  margin-bottom: 8px;
-}
-
-.dropdown .item:hover {
-  background-color: #eee;
-}
-
-.dropdown .item .text {
-  display: flex;
-  line-height: 20px;
-  flex-grow: 1;
-  min-width: 150px;
-}
-
-.dropdown .item .icon {
-  display: flex;
-  width: 20px;
-  height: 20px;
-  user-select: none;
-  margin-right: 12px;
-  line-height: 16px;
-  background-size: contain;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-
-.dropdown .divider {
-  width: auto;
-  background-color: #eee;
-  margin: 4px 8px;
-  height: 1px;
-}
-
-@media screen and (max-width: 1100px) {
-  .dropdown-button-text {
-    display: none !important;
-  }
-
-  .dialog-dropdown > .dropdown-button-text {
-    display: flex !important;
-  }
-
-  .font-size .dropdown-button-text {
-    display: flex !important;
-  }
-
-  .code-language .dropdown-button-text {
-    display: flex !important;
-  }
-}
-
-.toolbar .block-controls .dropdown-button-text {
-  width: 7em;
-  text-align: left;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  display: inline-block;
-}
-
-.dialog-dropdown {
-  background-color: #eee !important;
-  margin-bottom: 10px;
-  width: 100%;
-}
-
-button.item.dropdown-item-active {
-  background-color: rgba(223, 232, 250, 0.3);
-}
-
-button.item.dropdown-item-active i {
-  opacity: 1;
-}
