--- src/components/LexicalFocused/components/Dropdown/DropdownContext.tsx
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-export type DropDownContextType = {
-  registerItem: (ref: React.RefObject<HTMLButtonElement>) => void;
-};
-
-export const DropDownContext = createContext<DropDownContextType | null>(null);
