--- src/components/LexicalFocused/components/Dropdown/DropdownItem.tsx
@@ -1,43 +1,0 @@
-import { useContext, useEffect, useRef } from 'react';
-
-import { DropDownContext } from './DropdownContext';
-
-export function DropDownItem({
-  children,
-  className,
-  onClick,
-  title,
-}: {
-  children: React.ReactNode;
-  className: string;
-  onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
-  title?: string;
-}) {
-  const ref = useRef<HTMLButtonElement>(null);
-
-  const dropDownContext = useContext(DropDownContext);
-
-  if (dropDownContext === null) {
-    throw new Error('DropDownItem must be used within a DropDown');
-  }
-
-  const { registerItem } = dropDownContext;
-
-  useEffect(() => {
-    if (ref && ref.current) {
-      registerItem(ref);
-    }
-  }, [ref, registerItem]);
-
-  return (
-    <button
-      className={className}
-      onClick={onClick}
-      ref={ref}
-      title={title}
-      type='button'
-    >
-      {children}
-    </button>
-  );
-}
