+++ src/components/LexicalFocused/components/Dropdown/index.ts
@@ -1,4 +1,1 @@
 export * from './Dropdown';
-export * from './DropdownItems';
-export * from './DropdownItem';
-export * from './DropdownContext';
